+++ assets/components/productPage/productPageContentBlock/productPageContentBlock.scss
@@ -29,7 +29,7 @@
   .component-left-margin-section__content {
     border-color: gu-colour(header-border);
   }
-  .weekly-feature-list .weekly-feature-list__item {
+  .component-product-page-features .component-product-page-features__item {
     border-color: gu-colour(garnett-neutral-5);
   }
 }
